requirements.txt
pymongo>=4.6.0
pandas>=2.0.0
datasets>=2.18.0
python-dotenv>=1.0.0
requests>=2.28.0
beautifulsoup4>=4.12.0
pymupdf>=1.23.0
scikit-learn>=1.3.0
imbalanced-learn>=0.12.0
joblib>=1.3.0
xgboost>=2.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pytest>=8.0.0

# GCP Vertex AI
google-cloud-aiplatform>=1.38.0
langchain>=0.3.0
langchain-google-vertexai>=2.0.0
langchain-core>=0.3.0

# MCP (Model Context Protocol)
mcp>=1.0.0

